pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0", "wheel",]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "RSI"
version = "0.1.0"
description = "RSI = Room Screen Immersion."

packages = [
    { include = "rsi", from = "src" },
]

authors = ["RonFalafel <ronchick.weiss@gmail.com>",]
maintainers = ["Lari Liuhamo <lari.liuhamo+pypi@gmail.com>",]

include = []
license = "Proprietary"
readme = "README.md"

homepage = "https://pypi.org/project/RSI/"
repository = "https://github.com/RonFalafel/RSI"
documentation = "https://github.com/RonFalafel/RSI/tree/main/docs"

keywords = [
    "python3",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Typing :: Typed",
]


[tool.poetry.dependencies]
python = "^3.9.0"
mss = "^10.0.0"
numpy = "^2.0.0"
PySimpleGUI = "^5.0.0"
python-dotenv = "^1.0.1"
requests = "^2.32.0"
rtoml = "^0.11.0"
yeelight = "^0.7.14"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
types-requests = "^2.32.0.20241016"


[tool.poetry.group.linters]
optional = true


[tool.poetry.group.linters.dependencies]
ruff = "^0.7.3"


[tool.poetry.urls]
"Tracker" = "https://github.com/Diapolo10/project-name/issues"
"Changelog" = "https://github.com/Diapolo10/project-name/blob/main/CHANGELOG.md"


[tool.ruff]
lint.select = [
    "A",     # Builtins
    "ANN",   # Annotations
    "ARG",   # Unused arguments
    "ASYNC", # Asynchronous code
    "B",     # Bugbear
    "BLE",   # Blind except
    "C4",    # Comprehensions
    "C90",   # mccabe
    "COM",   # Commas
    "D1",    # Undocumented public elements
    "D2",    # Docstring conventions
    "D3",    # Triple double quotes
    "D4",    # Docstring text format
    "DTZ",   # Datetimes
    "E",     # Errors
    "EM",    # Error messages
    "ERA",   # Commented-out code
    "EXE",   # Executable
    "F",     # Pyflakes
    "FA",    # __future__ annotations
    "FBT",   # "Boolean trap"
    "FIX",   # "FIXME"-comments
    "FLY",   # F-strings
    "FURB",  # Refurb
    "G",     # Logging format
    "I",     # Isort
    "ICN",   # Import conventions
    "INP",   # Disallow PEP-420 (Implicit namespace packages)
    "INT",   # gettext
    "ISC",   # Implicit str concat
    "LOG",   # Logging
    "N",     # PEP-8 Naming
    "NPY",   # Numpy
    "PERF",  # Unnecessary performance costs
    "PGH",   # Pygrep hooks
    "PIE",   # Unnecessary code
    "PL",    # Pylint
    "PT",    # Pytest
    "PTH",   # Use Pathlib
    "PYI",   # Stub files
    "Q",     # Quotes
    "RET",   # Return
    "RUF",   # Ruff
    "RSE",   # Raise
    "S",     # Bandit
    "SIM",   # Code simplification
    "SLF",   # Private member access
    "SLOT",  # __slots__
    "T10",   # Debugger
    "T20",   # Print
    "TCH",   # Type checking
    "TD",    # "TODO"-comments
    "TID",   # Tidy imports
    "TRY",   # Exception handling
    "UP",    # Pyupgrade
    "W",     # Warnings
    "YTT",   # sys.version
]
lint.ignore = [
    "ANN101",  # Type annotation for `self`
    "D203",    # One blank line before class docstring
    "D212",    # Multi-line summary first line
    "PLR0913", # Too many arguments
    "Q000",    # Single quotes found but double quotes preferred
]
line-length = 120
show-fixes = true
src = ["src",]
target-version = "py39"


[tool.ruff.lint.flake8-copyright]
author = "RonFalafel"


[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
multiline-quotes = "double"


[tool.ruff.lint.mccabe]
max-complexity = 10


[tool.ruff.lint.per-file-ignores]
# https://beta.ruff.rs/docs/rules/
"__init__.py" = ["F401", "F403", "F405",]
"tests/*" = ["ANN", "ARG", "INP001", "S101",]
"logger.py" = ["N815",]


[tool.ruff.lint.pylint]
max-args = 15
max-branches = 20
max-returns = 10
max-statements = 80


[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
